pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-protocol==0.1.8",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anthropic==0.64.0",
    "anyio==4.10.0",
    "argcomplete==3.6.2",
    "attrs==25.3.0",
    "authlib==1.6.3",
    "boto3==1.40.21",
    "botocore==1.40.21",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cohere==5.17.0",
    "colorama==0.4.6",
    "cryptography==45.0.6",
    "cyclopts==3.23.0",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "docstring-parser==0.17.0",
    "docutils==0.22",
    "email-validator==2.3.0",
    "eval-type-backport==0.2.2",
    "exceptiongroup==1.3.0",
    "executing==2.2.0",
    "fastavro==1.12.0",
    "fastmcp==2.11.3",
    "filelock==3.19.1",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "genai-prices==0.0.24",
    "google-auth==2.40.3",
    "google-genai==1.32.0",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "griffe==1.13.0",
    "groq==0.31.0",
    "grpcio==1.74.0",
    "h11==0.16.0",
    "hf-xet==1.1.9",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.34.4",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "invoke==2.2.0",
    "ipykernel>=6.30.1",
    "isodate==0.7.2",
    "jiter==0.10.0",
    "jmespath==1.0.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.4.1",
    "jupyter>=1.1.1",
    "langchain==0.3.27",
    "langchain-community==0.3.29",
    "langchain-core==0.3.75",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.32",
    "langchain-text-splitters==0.3.10",
    "langgraph>=0.6.7",
    "langsmith==0.4.21",
    "lazy-object-proxy==1.12.0",
    "logfire==4.3.6",
    "logfire-api==4.3.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mcp==1.13.1",
    "mdurl==0.1.2",
    "milvus-lite==2.5.1",
    "mistralai==1.9.9",
    "more-itertools==10.7.0",
    "multidict==6.6.4",
    "mypy-extensions==1.1.0",
    "mysql-connector-python>=9.4.0",
    "nexus-rpc==1.1.0",
    "notebook>=7.4.5",
    "numpy==2.3.2",
    "openai==1.102.0",
    "openapi-core==0.19.5",
    "openapi-pydantic==0.5.1",
    "openapi-schema-validator==0.6.3",
    "openapi-spec-validator==0.7.2",
    "opentelemetry-api==1.36.0",
    "opentelemetry-exporter-otlp-proto-common==1.36.0",
    "opentelemetry-exporter-otlp-proto-http==1.36.0",
    "opentelemetry-instrumentation==0.57b0",
    "opentelemetry-proto==1.36.0",
    "opentelemetry-sdk==1.36.0",
    "opentelemetry-semantic-conventions==0.57b0",
    "orjson==3.11.3",
    "packaging==25.0",
    "pandas==2.3.2",
    "parse==1.20.2",
    "pathable==0.4.4",
    "prompt-toolkit==3.0.52",
    "propcache==0.3.2",
    "protobuf==5.29.5",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-ai==0.8.1",
    "pydantic-ai-slim==0.8.1",
    "pydantic-core==2.33.2",
    "pydantic-evals==0.8.1",
    "pydantic-graph==0.8.1",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pymilvus==2.6.1",
    "pymysql>=1.1.2",
    "pyperclip==1.9.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rich==14.1.0",
    "rich-rst==1.3.1",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "s3transfer==0.13.1",
    "setuptools==78.1.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "sse-starlette==3.0.2",
    "starlette==0.47.3",
    "temporalio==1.16.0",
    "tenacity==9.1.2",
    "tokenizers==0.22.0",
    "tqdm==4.67.1",
    "typer==0.16.1",
    "types-protobuf==6.30.2.20250822",
    "types-requests==2.32.4.20250809",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "ujson==5.11.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
    "werkzeug==3.1.1",
    "wheel==0.45.1",
    "wrapt==1.17.3",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.24.0",
]
